# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MetaWall API",
    "description": "MetaWall 生成文件"
  },
  "host": "localhost:3005",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "description": "取得全部貼文",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "type": "string",
            "required": false,
            "description": "排序：desc/asc"
          },
          {
            "name": "keyword",
            "in": "query",
            "type": "string",
            "required": false,
            "description": "關鍵字搜尋"
          }
        ],
        "responses": {
          "200": {
            "description": "貼文資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "取得成功"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "王小明"
                      },
                      "photo": {
                        "type": "string",
                        "example": "aaa.jpg"
                      },
                      "content": {
                        "type": "string",
                        "example": "這是我的貼文內容"
                      },
                      "image": {
                        "type": "string",
                        "example": "bbb.jpg"
                      },
                      "likes": {
                        "type": "number",
                        "example": 20000
                      },
                      "_id": {
                        "type": "string",
                        "example": "6276252a59d61d0687896756"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2022-05-07T07:52:10.937Z"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "description": "新增單一貼文",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "string",
                  "example": "626c14b5fdb48b9d205ef262"
                },
                "content": {
                  "type": "string",
                  "example": "這是我的貼文內容"
                },
                "image": {
                  "type": "string",
                  "example": "bbb.jpg"
                },
                "likes": {
                  "type": "number",
                  "example": 300
                }
              },
              "required": [
                "user",
                "content"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "貼文資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "取得成功"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "王小明"
                    },
                    "photo": {
                      "type": "string",
                      "example": "aaa.jpg"
                    },
                    "content": {
                      "type": "string",
                      "example": "這是我的貼文內容"
                    },
                    "image": {
                      "type": "string",
                      "example": "bbb.jpg"
                    },
                    "likes": {
                      "type": "number",
                      "example": 20000
                    },
                    "_id": {
                      "type": "string",
                      "example": "6276252a59d61d0687896756"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-05-07T07:52:10.937Z"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "description": "刪除全部貼文",
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "刪除成功"
                },
                "data": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}